Refactor create-file action: rename example.ts to create-file.ts, update file path handling, and add unit tests
