+++ plugins/scaffolder-backend-module-create-file/src/module.ts
@@ -1,6 +1,6 @@
-import { createBackendModule } from "@backstage/backend-plugin-api";
-import { scaffolderActionsExtensionPoint  } from '@backstage/plugin-scaffolder-node/alpha';
-import { createExampleAction } from "./actions/example";
+import { createBackendModule } from '@backstage/backend-plugin-api';
+import { scaffolderActionsExtensionPoint } from '@backstage/plugin-scaffolder-node/alpha';
+import { createExampleAction } from './actions/create-file';
 
 /**
  * A backend module that registers the action into the scaffolder
@@ -11,11 +11,11 @@ export const scaffolderModule = createBackendModule({
   register({ registerInit }) {
     registerInit({
       deps: {
-        scaffolderActions: scaffolderActionsExtensionPoint
+        scaffolderActions: scaffolderActionsExtensionPoint,
       },
-      async init({ scaffolderActions}) {
+      async init({ scaffolderActions }) {
         scaffolderActions.addActions(createExampleAction());
-      }
+      },
     });
   },
-})
+});
